refactor: PantryPage and added hook and service to add new groceries
